front: using RTK for get infra on reference map
